Automation_Ai_Scripts/ai_blog_scripts/requirements.txt
openai
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pandas
flask
# chatterbot
# chatterbot_corpus
ruamel.yaml>=0.17.21
pyjwt
python-dotenv
discord.py
schedule
joblib
scikit-learn
matplotlib
tweepy
textstat
numpy
pandas
language_tool_python
gunicorn
